+++ src/main/resources/org/patternfly/sample/SpinnerSample.kt
@@ -1,0 +1,14 @@
+package org.patternfly.sample
+
+import dev.fritz2.dom.html.render
+import org.patternfly.Size.MD
+import org.patternfly.spinner
+
+internal interface SpinnerSample {
+
+    fun spinner() {
+        render {
+            spinner(size = MD)
+        }
+    }
+}