+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.0]
 ### Added
 - Expandable cards ([#10](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/10))
+- Spinner component ([#51](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/51))  
 - Text component ([#53](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/53))
 - Title component ([#58](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/58))
 - Vertical divider ([#19](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/19))